Fix rmdir does not remove not empty directory
Remove SDFat V1 support as not necessary now have SDFat 2.x
Remove SPIFFS esp8266 support as now removed from core
Bump version
Remove outdated test scripts

+++ esp3d/src/modules/http/handles/upload-SD-files.cpp
@@ -118,7 +118,7 @@ void HTTP_Server::SDFileupload ()
                                     break;
                                 }
                             }
-                            if (pos+1 >= path.length()-1) {
+                            if ((size_t)(pos+1) >= path.length()-1) {
                                 pos=-1;
                                 break;
                             } else {
